+++ src/rosdiscover/observer/ros2.py
@@ -3,12 +3,13 @@ from roswire import AppInstance
 
 from .observer import Observer
 from .. import Config
+from ..interpreter import SystemSummary
 
 
 class ROS2Observer(Observer):
 
-    def __init__(self, app: AppInstance, config: Config):
+    def __init__(self, app: AppInstance, config: Config) -> None:
         super().__init__(app, config)
 
-    def observe_and_summarise(self):
+    def observe(self) -> SystemSummary:
         raise NotImplementedError
